Erste Version von MS (2.7.2014): Jochen,
Attached you will find our current documentation. Note that this documentation is currently in flux and is subject to change. We are currently solidifying our ‘GET’ scenarios and then will move on to the ‘POST’ scenarios. We are also investigating our authentication model. I will update you once this is complete but it looks like we will be using Live Connect (as the NNTP Bridge does today) to handle user tokens. There are some dependencies which are blocking us from using OAuth.

If you could please review the current documentation and let us know if you see any gaps from your perspective. Please also ignore the versioning and the URL since these treatments are not locked.

Thank you,

Bryant Hawthorne