Neue Version (14.7.2014):
Hey Jochen!

Thank you for taking the time to go over this! I was meaning to reach out to you again since we made a few tweaks to our approach late last week (regarding how replies are surfaced, this should some of your second bullet). I have attached our current documentation. You do bring up a good point regarding the additional calls to get the users display name (I’m assuming this would be minimal viable?). I will review this with the team to see if this is something we can get in for V1. Is this something that would be blocking you if it does not make it?

For OAuth, I agree with you that that it is an important feature. However, we have several dependencies right now that are blocking us from updating to OAuth. We are predicting that within the year these dependencies will be resolved and we can then update our authentication both through the UI and our APIs.

For ‘brands’ and ‘categories’ you are correct. A forum can belong to multiple categories and even multiple brands. This is the same as the webUI. Take the feedback forum for example. This lives in TechNet, MSDN and the Microsoft brand:

http://social.msdn.microsoft.com/Forums/en-US/home?forum=suggest&filter=alltypes&sort=lastpostdesc
http://social.technet.microsoft.com/Forums/en-US/home?forum=suggest&filter=alltypes&sort=lastpostdesc
http://social.microsoft.com/Forums/en-US/home?forum=suggest&filter=alltypes&sort=lastpostdesc

Thanks again Jochen! Let me know if there are any other issues you can see.

Thanks,

Bryant Hawthorne